colpali/llm_vision_model_host/requirements.txt
fastapi
uvicorn
Pillow
torch
transformers
accelerate
huggingface_hub
requests
pydantic